feature: 敲击信号发现算法

+++ app/src/main/java/tech/oom/idealrecorderdemo/dto/ModelInstDebugDto.java
@@ -6,9 +6,9 @@ public class ModelInstDebugDto {
     public int dataset_id;
     public String algorithm;
     public String label;
-    public List<Short> signal;
+    public List<Integer> signal;
 
-    public ModelInstDebugDto(int dataset_id, String algorithm, String label, List<Short> signal) {
+    public ModelInstDebugDto(int dataset_id, String algorithm, String label, List<Integer> signal) {
         this.dataset_id = dataset_id;
         this.algorithm = algorithm;
         this.label = label;
